# Remove unnecessary allocations