@robert-milan 2019-07-11
I will do the benchmarking this weekend if possible with some of these suggestions (atomics in certain places, re-using AddOrGet / Delete where possible, branches, etc...). I'll implement whatever ends up performing best and then merge.