@robert-milan 2019-07-12
After some more thought I have realized that I cannot use atomics and need the locks. They are also protecting the index. Without the locks there is a chance that an object will be deleted after an `add` method has acquired its memory address. Then when the `add` method tries to `increment` the `reference counter` it will be writing to memory it should not have access to.